+++ src/euromail/__init__.py
@@ -2,10 +2,15 @@
 
 from euromail.client import EuroMail
 from euromail.async_client import AsyncEuroMail
+from euromail.webhooks import verify_signature, DEFAULT_TOLERANCE_SECONDS
 from euromail.errors import (
     EuroMailError,
     AuthenticationError,
+    ConflictError,
+    ForbiddenError,
+    NotFoundError,
     RateLimitError,
+    ServerError,
     ValidationError,
 )
 from euromail.types import (
@@ -34,6 +39,7 @@ from euromail.types import (
     EmailValidation,
     GdprEraseResult,
     GdprExport,
+    ImportSuppressionsResult,
     InboundEmail,
     InboundRoute,
     Newsletter,
@@ -70,9 +76,15 @@ from euromail.types import (
 __all__ = [
     "EuroMail",
     "AsyncEuroMail",
+    "verify_signature",
+    "DEFAULT_TOLERANCE_SECONDS",
     "EuroMailError",
     "AuthenticationError",
+    "ConflictError",
+    "ForbiddenError",
+    "NotFoundError",
     "RateLimitError",
+    "ServerError",
     "ValidationError",
     "Account",
     "AnalyticsPeriod",
@@ -99,6 +111,7 @@ __all__ = [
     "EmailValidation",
     "GdprEraseResult",
     "GdprExport",
+    "ImportSuppressionsResult",
     "InboundEmail",
     "InboundRoute",
     "Newsletter",